refactor(event-sender): remove unecessary folder structure to simplify worker
