+++ src/event-sender/src/modules/configs/app-config.ts
