+++ src/event-sender/src/modules/dependencies/create-dependencies-container.ts
