+++ src/event-sender/src/modules/repositories/interfaces/event.repository.ts
